♻️ move power utilities and native methods to separate folder
